requirements.txt
asgiref==3.2.8
Django==3.0.7
django-crispy-forms==1.9.0
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2-binary==2.8.4
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
